product update done.

+++ app/Http/Controllers/ProductController.php
@@ -9,7 +9,6 @@ use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Storage;
-use Symfony\Component\Console\Input\Input;
 use File;
 
 class ProductController extends Controller
@@ -18,31 +17,41 @@ class ProductController extends Controller
     /**
      * Display a listing of the resource.
      *
-     * @return Response
+     * @return view
      */
     public function index()
     {
-        $products = Product::all();
-        return view('product.index')->with([
-           'products' => $products
-        ]);
+        if(auth()->user()->can('index product')){
+            $products = Product::all();
+            return view('product.index')->with([
+                'products' => $products
+            ]);
+        }else{
+            return redirect()->route('home')->with('error','Unauthorized access!');
+        }
+
     }
 
     /**
      * Show the form for creating a new resource.
      *
-     * @return Response
+     * @return view
      */
     public function create()
     {
-        $categories = Category::all();
-        $subcategories = SubCategory::all();
-        $product = Product::find(16);
-        return view('product.create')->with([
-            'categories' => $categories,
-            'subcategories' => $subcategories,
-            'product' => $product
-        ]);
+        if(auth()->user()->can('create product')){
+            $categories = Category::all();
+            $subcategories = SubCategory::all();
+            $product = Product::find(16);
+            return view('product.create')->with([
+                'categories' => $categories,
+                'subcategories' => $subcategories,
+                'product' => $product
+            ]);
+        }else{
+            return redirect()->route('home')->with('error','Unauthorized access!');
+        }
+
     }
 
 
@@ -54,57 +63,60 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        $request->validate([
-            'title' => 'required|string|max:20',
-            'description' => 'required|string',
-            'price' => 'required|numeric',
-            'salePrice' => 'nullable|numeric',
-            'images.*' => 'nullable|image|dimensions:ratio=12/13,min_width=600,min_height=650,max_width=1200,max_height=1300'
-        ]);
-        $product = new Product;
-        $product->title = $request->input('title');
-        $product->description = $request->input('description');
-        $product->price = $request->input('price');
-        if($request->has('sale_price')){
-            $product->discounted_price = $request->input('sale_price');
-        }
-        if($request->has('sku')){
-            $product->sku = $request->input('sku');
-        }
-        if($request->input('in_stock') === 'instock'){
-            $product->in_stock = true;
+        if(auth()->user()->can('store product')){
+            $request->validate([
+                'title' => 'required|string|max:20',
+                'description' => 'required|string',
+                'price' => 'required|numeric',
+                'salePrice' => 'nullable|numeric',
+                'images.*' => 'nullable|image|dimensions:ratio=12/13,min_width=600,min_height=650,max_width=1200,max_height=1300'
+            ]);
+            $product = new Product;
+            $product->title = $request->input('title');
+            $product->description = $request->input('description');
+            $product->price = $request->input('price');
+            if($request->has('sale_price')){
+                $product->discounted_price = $request->input('sale_price');
+            }
+            if($request->has('sku')){
+                $product->sku = $request->input('sku');
+            }
+            if($request->input('in_stock') === 'instock'){
+                $product->in_stock = true;
+            }else{
+                $product->in_stock = false;
+            }
+            if($request->has('quantity')){
+                $product->quantity = $request->input('quantity');
+            }
+            if($request->has('purchase_note')){
+                $product->purchase_note = $request->input('purchase_note');
+            }
+            $product->is_active = $request->has('is_active');
+            $product->is_featured = $request->has('is_featured');
+            try{
+                $product->save();
+                if($request->has('category_id')){
+                    foreach ($request->input('category_id') as $category_id){
+                        $category = Category::find($category_id);
+                        $product->categories()->attach($category);
+                    }
+                }
+                if($request->has('images')){
+                    foreach ($request->file('images') as $img){
+
+                        $product->images()->create([
+                            'url' =>  Storage::url($img->store('public/images/products'))
+                        ]);
+                    }
+                }
+                return redirect()->back()->with('success','stored successfully');
+            }catch (\Exception $e){
+                return redirect()->back()->with('error',$e->getCode().':'.$e->getMessage());
+            }
         }else{
-            $product->in_stock = false;
+            return redirect()->route('home')->with('error','Unauthorized access!');
         }
-        if($request->has('quantity')){
-            $product->quantity = $request->input('quantity');
-        }
-        if($request->has('purchase_note')){
-            $product->purchase_note = $request->input('purchase_note');
-        }
-        $product->is_active = $request->has('is_active');
-        $product->is_featured = $request->has('is_featured');
-        try{
-            $product->save();
-            if($request->has('category_id')){
-                foreach ($request->input('category_id') as $category_id){
-                    $category = Category::find($category_id);
-                    $product->categories()->attach($category);
-                }
-            }
-            if($request->has('images')){
-                foreach ($request->file('images') as $img){
-
-                    $product->images()->create([
-                       'url' =>  Storage::url($img->store('public/images/products'))
-                    ]);
-                }
-            }
-            return redirect()->back()->with('success','stored successfully');
-        }catch (\Exception $e){
-            return redirect()->back()->with('error',$e->getCode().':'.$e->getMessage());
-        }
-
     }
 
 
@@ -116,12 +128,24 @@ class ProductController extends Controller
      */
     public function edit($id)
     {
-        $product = Product::find($id);
-        $categories = Category::all();
-        return view('product.edit')->with([
-           'product' => $product,
-            'categories' => $categories
-        ]);
+        if(auth()->user()->can('edit product')) {
+            if (is_numeric($id)) {
+                $product = product::find($id);
+                if ($product == null) {
+                    return redirect()->back()->with('error', 'product not exists!');
+                }
+                $categories = Category::all();
+                return view('product.edit')->with([
+                    'product' => $product,
+                    'categories' => $categories
+                ]);
+            }else{
+                return redirect()->back()->withErrors('wrong url!');
+            }
+        }else{
+            return redirect()->route('home')->with('error','Unauthorized access!');
+        }
+
     }
 
     /**
@@ -134,25 +158,29 @@ class ProductController extends Controller
      */
     public function destroy_image($product_id, $image_id)
     {
-        $response = [];
-        try{
-            $product = Product::find($product_id);
-            $image = Image::find($image_id);
-            $response['data'] = $image;
-            if($product->images->contains($image)){
-                if(File::exists(public_path($image->url))){
-                    File::delete(public_path($image->url));
+        if(auth()->user()->can('edit product')){
+            $response = [];
+            try{
+                $product = Product::find($product_id);
+                $image = Image::find($image_id);
+                $response['data'] = $image;
+                if($product->images->contains($image)){
+                    if(File::exists(public_path($image->url))){
+                        File::delete(public_path($image->url));
+                    }else{
+                        $response['message'] = 'image does not exists';
+                    }
                 }else{
-                    $response['message'] = 'image does not exists';
+                    $response['message'] = 'this image does not belongs to the specified product';
                 }
-            }else{
-                $response['message'] = 'this image does not belongs to the specified product';
+                $product->images()->detach($image_id);
+                $image->delete();
+                $response['message'] = 'Successfully deleted this image';
+            }catch(\Exception $e){
+                $response['message'] = $e->getMessage();
             }
-            $product->images()->detach($image_id);
-            $image->delete();
-            $response['message'] = 'Successfully deleted this image';
-        }catch(\Exception $e){
-            $response['message'] = $e->getMessage();
+        }else{
+            $response['message'] = 'Unauthorized access!';
         }
         return $response;
     }
@@ -165,7 +193,63 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //
+        if(auth()->user()->can('edit product')){
+            $request->validate([
+                'title' => 'required|string|max:20',
+                'description' => 'required|string',
+                'price' => 'required|numeric',
+                'salePrice' => 'nullable|numeric',
+                'images.*' => 'nullable|image|dimensions:ratio=12/13,min_width=600,min_height=650,max_width=1200,max_height=1300'
+            ]);
+            if(is_numeric($id)){
+                $product =Product::find($id);
+                if($product == null){
+                    return redirect()->back()->with('error','product does not exists!');
+                }
+                $product->title = $request->input('title');
+                $product->description = $request->input('description');
+                $product->price = $request->input('price');
+                if($request->has('sale_price')){
+                    $product->discounted_price = $request->input('sale_price');
+                }
+                if($request->has('sku')){
+                    $product->sku = $request->input('sku');
+                }
+                if($request->input('in_stock') === 'instock'){
+                    $product->in_stock = true;
+                }else{
+                    $product->in_stock = false;
+                }
+                if($request->has('quantity')){
+                    $product->quantity = $request->input('quantity');
+                }
+                if($request->has('purchase_note')){
+                    $product->purchase_note = $request->input('purchase_note');
+                }
+                $product->is_active = $request->has('is_active');
+                $product->is_featured = $request->has('is_featured');
+                try{
+                    //
+                    $product->save();
+                    if($request->has('category_id')){
+                        $product->categories()->sync($request->input('category_id'));
+                    }
+                    if($request->has('images')){
+                        foreach ($request->file('images') as $img){
+
+                            $product->images()->create([
+                                'url' =>  Storage::url($img->store('public/images/products'))
+                            ]);
+                        }
+                    }
+                    return redirect()->route('product.index')->with('success','updated successfully');
+                }catch (\Exception $e){
+                    return redirect()->back()->with('error',$e->getCode().':'.$e->getMessage());
+                }
+            }
+        }else{
+            return redirect()->route('home')->with('error','Unauthorized access!');
+        }
     }
 
     /**
@@ -176,19 +260,24 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        $product = Product::find($id);
-        try{
-            foreach($product->images()->get() as $img){
-                if(File::exists(public_path($img->url))){
-                    File::delete(public_path($img->url));
+        if(auth()->user()->can('delete product')){
+            $product = Product::find($id);
+            try{
+                foreach($product->images()->get() as $img){
+                    if(File::exists(public_path($img->url))){
+                        File::delete(public_path($img->url));
+                    }
                 }
+                $product->images()->detach();
+                $product->categories()->detach();
+                $product->delete();
+                return redirect()->route('product.index')->with('success','successfully deleted');
+            }catch(\Exception $e){
+                return redirect()->route('product.index')->withErrors($e->getMessage());
             }
-            $product->images()->detach();
-            $product->categories()->detach();
-            $product->delete();
-            return redirect()->route('product.index')->with('success','successfully deleted');
-        }catch(\Exception $e){
-            return redirect()->route('product.index')->withErrors($e->getMessage());
+        }else{
+            return redirect()->route('home')->with('error','Unauthorized access!');
         }
+
     }
 }
